Batch Input

Batch Input

GeoEfficiency.typeofSrc — Function.

typeofSrc()::SrcType

return the current value of the global GeoEfficiency.srcType.

source

typeofSrc(x::Int)::SrcType

set and return the value of the global GeoEfficiency.srcType corresponding to x.

srcUnknown = -1 also any negative integer treated as so,

srcPoint = 0,

srcLine = 1,

srcDisk = 2,

srcVolume = 3,

srcNotPoint = 4 also any greater than 4 integer treated as so.

source

GeoEfficiency.setSrcToPoint — Function.

setSrcToPoint()::Bool

return whether the source type is a point or not.

source

setSrcToPoint(yes::Bool)::Bool

return whether the source type is a point or not after setting srcType to srcPoint if yes = true else if yes = false setting it to srcNotPoint if it was not already set to other non-point type (srcDisk, srcLine, srcVolume).

Note

The user can use this function to change the source type any time.

The source type is set the fist time asked for source.

see also: typeofSrc(::Int).

source

setSrcToPoint(prompt::AbstractString)::Bool

return whether the source type is a point or not. only prompt the user to set the source type if it were not already set before.

see also: typeofSrc(::Int), setSrcToPoint(::Bool).

source

GeoEfficiency.detector_info_from_csvFile — Function.

UnExported

 detector_info_from_csvFile(detectors::AbstractString = Detectors, 
                                  datadir::AbstractString = dataDir)

return a vector{Detector} based on information in the file of name detectors found in the directory datadir.

Note

if no path is given the second argument datadir is default to C:\Users\Mohamed\GeoEfficiency as set by the constant dataDir.

if no file name is specified the name of the predefined file Detectors.csv as set by the constant Detectors.

the no argument method is the most useful; other methods are mainly for test propose.

source

GeoEfficiency.read_from_csvFile — Function.

UnExported

read_from_csvFile(csv_data::AbstractString, 
                   datadir::AbstractString = dataDir)::Vector{Float64}

return Vector{Float64} based on data in csv file named csv_data. directory datadir point to where the file is located default to $C:\Users\Mohamed\GeoEfficiency$ as set by the constant dataDir.

source

GeoEfficiency.read_batch_info — Function.

UnExported

read_batch_info()

read detectors and sources parameters from the predefined csv files.

Return a tuple (detectorsarray, srcHeightsarray, srcRhosarray, srcRadiiarray, srcLengthsarray, GeoEfficiencyisPoint)

source

UnExported

read_batch_info(datadir::AbstractString,
              detectors::AbstractString, 
             srcHeights::AbstractString,
                srcRhos::AbstractString,
               srcRadii::AbstractString,
             srcLengths::AbstractString)

read detectors and sources parameters from the location given in the argument list.

Return a tuple

   (detectors_array,
	srcHeights_array,
	srcRhos_array,
	srcRadii_array,
	srcLengths_array,
	isPoint)

source

GeoEfficiency.getDetectors — Function.

getDetectors(detectors_array::Vector{<:Detector} = Detector[])::Vector{Detector}

return the detectors_array as Vector{Detector} extended by the entered detectors and sorted according to the detector volume. prompt the user to input detector parameters from the console.

Note

If no array received in the input an empty array will be created to receive the converted detectors.

source

getDetectors(detector_info_array::Matrix{<:Real}, 
				 detectors_array::Vector{<:Detector} = Detector[]; 
				 						   console_FB=true)::Vector{Detector}

return detectors_array as Vector{Detector}, after extending it with the successfully converted detectors. while, attempt to convert detectors from the information in detector_info_array.

Note

if console_FB argument is set to true , the function will call getDetectors() to take input from the console if the detector_info_array is empty or contain no numerical element.

source