Error

Error

GeoEfficiency.GeoException — Type.

coustom abstract Exception that is the parent of all exception in the GeoEfficiency package

source

GeoEfficiency.InValidDetectorDim — Type.

coustom Exception indicating inValid radiation detector dimentions

source

GeoEfficiency.@validateDetector — Macro.

@validateDetector cond [text]

Throw an InValidDetectorDim if cond is false. Message text is optionally displayed upon validation failure.

Examples

julia> @validateDetector iseven(3) "3 is an odd number!"
ERROR: InValidDetectorDim: 3 is an odd number!

julia> @validateDetector isodd(3) "What even are numbers?"

source

GeoEfficiency.NotImplementedError — Type.

coustom Exception source to detector condation which may be valid but not implemented yet

source

GeoEfficiency.@notImplementedError — Macro.

coustom macro to throw NotImplementedError Exception

source