Physics Model

Physics Model

GeoEfficiency.Point — Type.

Point(Height::Real, Rho::Real)

construct and return a Point source that can be used as either a source by itself or an anchor point of a higher dimension source.

Height : point height relative to the detector surface.

Rho : point off-axis relative to the detector axis of symmetry.

Note

Each detector type give different interpretation to the height as follow:-

for CylDetector the point source height is consider to be measured from the detector face surface.

for BoreDetector the point source height is consider to be measured from the detector middle, +ve value are above the detector center while -ve are below.

for WellDetector the point source height is considered to be measured from the detector hole surface.

source

GeoEfficiency.source — Function.

source(anchorPnt::Point = Point())

return a tuple that describe the source (anchorPnt, SrcRadius, SrcLength) according to the input from the console.

anchorPnt : the source anchoring point. if it is missing the user is prompt to input it via the console.

SrcRadius : source radius.

SrcLength : source length.

Warning

If the global variable srcType is set to $srcPoint$, both SrcRadius and SrcLength are set to zero.

source

GeoEfficiency.CylDetector — Type.

CylDetector(CryRadius::Real, CryLength::Real)

construct and return a cylindrical detector of the given crystal dimensions:-

CryRadius : the detector crystal radius.

CryLength : the detector crystal length.

Warning

both CryRadius and CryLength should be positive, while CryLength can also be set to zero.

source

GeoEfficiency.BoreDetector — Type.

BoreDetector(CryRadius::Real, CryLength::Real, HoleRadius::Real)

construct and return a bore-hole detector of the given crystal dimensions:-

CryRadius : the detector crystal radius.

CryLength : the detector crystal length.

HoleRadius : the detector hole radius.

Warning

CryRadius and CryLength, HoleRadius should be positive numbers, also CryRadius should be greater than HoleRadius.

source

GeoEfficiency.WellDetector — Type.

WellDetector(CryRadius::Real, CryLength::Real, HoleRadius::Real, HoleDepth::Real)

construct and return a Well-Type detector of the given crystal dimensions:-

CryRadius : the detector crystal radius.

CryLength : the detector crystal length.

HoleRadius : the detector hole radius.

HoleDepth : the detector hole length.

Warning

all arguments should be positive numbers, also CryRadius should be greater than HoleRadius and CryLength should be greater than HoleDepth.

source

GeoEfficiency.RadiationDetector — Type.

abstract super-supertype of all detectors types

source

GeoEfficiency.Detector — Type.

Detector

abstract supertype of all detectors types of cylidericalish shapes. also can be used to construct any leaf type.

source