Summery

Summery

GeoEfficiency.about — Function.

 *************************************************
 **            -=) GeoEfficiency (=-             **
 **  Accurate Geometrical Efficiency Calculator  **
 **   First Created on Fri Aug 14 20:12:01 2015  **
 *************************************************

 Author:        Mohamed E. Krar,  @e-mail: DrKrar@gmail.com 
 Auth_Profile:  https://www.researchgate.net/profile/Mohamed_Krar3
 Repository:    https://github.com/DrKrar/GeoEfficiency.jl/
 Version:       v"0.9.4-dev" - (0 days old master)  
 Documentation: https://GeoEfficiency.GitHub.io/dev/index.html
 PDF_Manual:    https://GeoEfficiency.GitHub.io/dev/pdf/GeoEfficiency.jl.pdf

Batch mode 
-  read files by defaul from directory `/home/GeoEfficiency`
-  save results by default to directory `/home/GeoEfficiency/results`

for more information see `batch`, `batchInfo`.

source

GeoEfficiency.GeoEfficiency — Module.

GeoEfficiency Package

introduce a fast and flexible tool to calculate in batch or individually the geometrical efficiency for a set of common radiation detectors shapes (cylindrical,Bore-hole, Well-type) as seen form a source. The source can be a point, a disc or even a cylinder.

Quick Usage

geoEff() : Calculate the geometrical efficiency for one geometrical setup return only the value of the geometrical efficiency.

calc() : Calculate the geometrical efficiency for one geometrical setup and display full information on the console.

calcN() : Calculate the geometrical efficiency for geometrical setup(s) and display full information on the console until the user quit.

batch() : Calculate the geometrical efficiency using data in the /home/GeoEfficiency folder in batch mode.

for more information and updates refer to the repository at GitHub.com

source