Physics Model

Physics Model

Geometrical efficiency of radioactive source measurement is a type of detection efficiency. A fully describe a radioactive source measurement at the most basic level three component should be provided.

radioactive detector description

radiation source description

relative position of the source to detector.

this section will discus how to instruct the program to construct each of the aforementioned component.

Detector

Currently, only cylindrical-like types of detectors are supported.

Cylindrical Detector

GeoEfficiency.CylDetector — Type.

CylDetector(CryRadius::Real, CryLength::Real)

construct and return a cylindrical detector of the given crystal dimensions:-

CryRadius : the detector crystal radius.

CryLength : the detector crystal length.

Warning

both CryRadius and CryLength should be positive, while CryLength can also be set to zero.

source

GeoEfficiency.CylDetector — Method.

CylDetector(CryRadius::Real)

construct and return a cylindrical (really disk) detector with crystal length equal to zero.

see also: CylDetector(CryRadius::Real, CryLength::Real).

source

GeoEfficiency.CylDetector — Method.

CylDetector()

construct and return a cylindrical detector according to the input from the console.

see also: CylDetector(CryRadius::Real, CryLength::Real).

source

Note

the positon of the source is reported relative to the detector anchoring point, for a cylinder detector it is taking as a point in the plain surface nearest to the source which lies on the detector axis of symmetry.

Bore-hole Detector

GeoEfficiency.BoreDetector — Type.

BoreDetector(CryRadius::Real, CryLength::Real, HoleRadius::Real)

construct and return a bore-hole detector of the given crystal dimensions:-

CryRadius : the detector crystal radius.

CryLength : the detector crystal length.

HoleRadius : the detector hole radius.

Warning

CryRadius and CryLength, HoleRadius should be positive numbers, also CryRadius should be greater than HoleRadius.

source

GeoEfficiency.BoreDetector — Method.

BoreDetector()

construct and return a bore-hole detector according to the input from the console.

see also: BoreDetector(CryRadius::Real, CryLength::Real, HoleRadius::Real).

source

Note

the positon of the source is reported relative to the detector anchoring point, for a bore-hole detector it is taking as the middle point of its axis of symmetry.

Well-type Detector

GeoEfficiency.WellDetector — Type.

WellDetector(CryRadius::Real, CryLength::Real, HoleRadius::Real, HoleDepth::Real)

construct and return a Well-Type detector of the given crystal dimensions:-

CryRadius : the detector crystal radius.

CryLength : the detector crystal length.

HoleRadius : the detector hole radius.

HoleDepth : the detector hole length.

Warning

all arguments should be positive numbers, also CryRadius should be greater than HoleRadius and CryLength should be greater than HoleDepth.

source

GeoEfficiency.WellDetector — Method.

WellDetector()

construct and return a Well-Type detector according to the input from the console.

see also: WellDetector(CryRadius::Real, CryLength::Real, HoleRadius::Real, HoleDepth::Real).

source

Note

the positon of the source is reported relative to the detector anchoring point, for well-type detector it is taking as the point detector hole surface that lies on the detector axis of symmetry.

Source

GeoEfficiency.source — Function.

source(anchorPnt::Point = Point())

return a tuple that describe the source (anchorPnt, SrcRadius, SrcLength) according to the input from the console.

anchorPnt : the source anchoring point. if it is missing the user is prompt to input it via the console.

SrcRadius : source radius.

SrcLength : source length.

Warning

if source type set to point source, both `SrcRadius` and `SrcLength` are set to zero. 
for more information **see also:** [`typeofSrc()`](@ref) and [`typeofSrc(x::Int)`](@ref).

source

Source Anchoring Point

GeoEfficiency.Point — Type.

Point(Height::Real, Rho::Real)

construct and return a Point source. The Point can be used as either a source by itself or an anchor point of a higher dimension source.

Height : point height relative to the detector surface.

Rho : point off-axis relative to the detector axis of symmetry.

Note

Each detector type give different interpretation to the height as follow:-

for CylDetector the point source height is consider to be measured from the detector face surface.

for BoreDetector the point source height is consider to be measured from the detector middle, +ve value are above the detector center while -ve are below.

for WellDetector the point source height is considered to be measured from the detector hole surface.

source

GeoEfficiency.Point — Method.

Point(Height::Real)

construct and return an axial point.

see also: Point(Height::Real, Rho::Real).

source

GeoEfficiency.Point — Method.

Point()

construct and return a point. prompt to input information via the console.

see also: Point(Height::Real, Rho::Real).

source

GeoEfficiency.Point — Method.

Point(xHeight::Real, aPnt::Point)

construct and return a point that has the same off-axis distance as aPnt but of new height xHeight.

see also: Point(Height::Real, Rho::Real)

source

GeoEfficiency.Point — Method.

Point(aPnt::Point, xRho::Real)

construct and return a point that has the same height as aPnt but of new off-axis distance Rho.

see also: Point(Height::Real, Rho::Real).

source